requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.5.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
pytz==2023.3
requests==2.31.0
telegraph==2.2.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
